% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_A8_A0A7_maf_data.R
\docType{data}
\name{TCGA_A8_A0A7_maf}
\alias{TCGA_A8_A0A7_maf}
\title{SNVsReadCountsFile for sample TCGA-A8-A0A7}
\format{A data.frame object.}
\description{
A data frame object containing the read counts data of somatic
single nucleotide variants (SNVs) loci for sample TCGA-A8-A0A7.
The header contains the chromosme that harbors the SNV ("chr" column),
the position of the SNV (defined by the "start" and "end" columns),
the informations about the reference and alternative base counts ("ref.count"
and	"alt.count" columns, respectively) and finally the sample ID ("sample"
column). For more information please visit
\href{https://software.broadinstitute.org/software/igv/MutationAnnotationFormat}{MAF file format}.
}
