% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_A8_A0A7_ploidy_data.R
\docType{data}
\name{TCGA_A8_A0A7_ploidy}
\alias{TCGA_A8_A0A7_ploidy}
\title{Ploidy data for sample TCGA-A8-A0A7}
\format{A data.frame object.}
\description{
A data frame containing the ploidy status of a sample. It must
contains at least the sample ID ("sample" column) and the ploidy status ("ploidy" column).
}
