% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_HT_8564_ploidy_data.R
\docType{data}
\name{TCGA_HT_8564_ploidy}
\alias{TCGA_HT_8564_ploidy}
\title{Ploidy data for sample TCGA-HT-8564}
\format{A data.frame object.}
\description{
A data frame containing the ploidy status of a sample. It must
contains at least the sample ID ("sample" column) and the ploidy status ("ploidy" column).
}
