% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_HT_8564_seg_data.R
\docType{data}
\name{TCGA_HT_8564_seg}
\alias{TCGA_HT_8564_seg}
\title{SEG file (segmented data) for sample TCGA-HT-8564}
\format{A data.frame object.}
\description{
A data frame object that lists loci and associated numeric values.
The header must be compatible with the standard format defined by the Broad
Institute. For more information please visit
\href{https://software.broadinstitute.org/software/igv/SEG}{SEG file format}.
}
